requirements.txt
requests==2.32.5
Flask==3.1.3
python-dotenv==1.2.2
gunicorn==25.1.0
Flask-WTF==1.2.2